Add retry to openai (#14)
Reviewer: @poornimaramesh 

Estimate: 5 mins

---

### Goal
Add retries to the OpenAI API call. Sometimes the API call returns
nothing which breaks the pipeline.

### Changes
- Added a retry decorator from the tenacity library.
- Renamed the _ask_llm_json to ask_llm_json (Removed _ since it was
being used outside the module)

## How has this been tested?
Difficult to test other than to fake the API endpoint but that will
obviously work.

This might not solve issues. But still good to have.

## To-do before merge
Run the test notebook and ensure nothing breaks.

## Checklist

Fill with `x` for completed. Delete any lines that are not relevant

- [x] My code follows the style guidelines of this project
- [x] I have reviewed my own code to ensure good quality
- [x] I have tested the functionality of my code to ensure it works as
intended
- [x] I have resolved merge conflicts
- [x] I have updated the automated tests (if applicable)
- [x] I have updated the requirements (if applicable)
- [ ] I have updated the README file (if applicable)
- [ ] I have updated affected documentation (if applicable)

---------

Co-authored-by: poornimaramesh <poornimaramesh1995@gmail.com>