Any LiteLLM model works and can be set up dynamically (#19)
Reviewer: @poornimaramesh

Estimate: 30 mins max

---

### Goal
Make it so that litellm calls to any provider or model can be set
dynamically through function params instead of env vars.

This let's us have different models for different users (goal).

### Changes
Added the llm_connection_config param to multiple functions (it is only
relevant for the ask_llm_json function but since it is deeply nested I
had to pass it through many intermediate layers)

## How has this been tested?
Ran the demo notebook and the validation notebook

## Checklist

Fill with `x` for completed. Delete any lines that are not relevant

- [x] My code follows the style guidelines of this project
- [x] I have reviewed my own code to ensure good quality
- [x] I have tested the functionality of my code to ensure it works as
intended
- [x] I have resolved merge conflicts
- [x] I have updated the README file (if applicable)